backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
numpy==1.21.2